SORUNSUZ CALISIR HALDE

+++ app/src/main/java/com/example/e_marketapp/util/MapperFunction.kt
@@ -10,7 +10,7 @@ fun baseModelToMarketBasketEntity(baseModelItem: BaseModelItem): MarketBasketEnt
         productId = baseModelItem.id,
         productName = baseModelItem.name,
         singleItemPrice = baseModelItem.price,
-        productPrice = baseModelItem.price.toDouble(),
+        productPrice = baseModelItem.price,
         productCount = 1
     )
 }
